feat(utils): new structure + rename willPaginate to Paginator
