+++ lib/resolvers/publicUser.ts
@@ -1,9 +1,9 @@
-import { safeApiCall, willPaginateFactoryFromVariables } from '../utils';
+import { safeApiCall, paginatorFromVariables } from '../utils';
 
 export function publicUserResolvers(spotifyApiClient) {
   return {
     playlists(user, variables) {
-      return willPaginateFactoryFromVariables('OffsetPaging', variables)(
+      return paginatorFromVariables('OffsetPaging', variables)(
         spotifyApiClient,
         'getUserPlaylists',
         (response) => response.body.items,
